requirement.txt
numpy==2.1.3
opencv-python==4.10.0.84
python-dotenv==1.0.1
